CUDA: Use pinned memory only for x and p
The other allocations (e.g. the matrix) hasn't worked as expected because
the virtual methods are currently called from the constructors. Hence, the
Matrix objects are not yet fully constructed and call the fallback functions
which leads to pageable memory.
When fixing this, the Total Runtime increases, so remove this entirely. Then,
we cannot use cudaMemcpyAsync because this requires pinned memory. Funnily
enough, this even improves the 'transfer to' time...